+++ tmp/lanelet2_extension/test/src/autoware_lanelet2_validation/test_missing_regulatory_elements.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "lanelet2_extension/autoware_lanelet2_validation/validators/missing_regulatory_elements.hpp"
+#include "lanelet2_extension/autoware_lanelet2_validation/vals/missing_regulatory_elements.hpp"
 
 #include <gtest/gtest.h>
